Cleaned up things for Pull Request - mostly stable! Still need to comment Value Iteration and visualization code
